+++ src/main/java/today/meevote/domain/schedule/dao/ScheduleDao.java
@@ -1,5 +1,6 @@
 package today.meevote.domain.schedule.dao;
 
+import java.util.List;
 import java.util.Map;
 
 import org.apache.ibatis.annotations.Mapper;
@@ -7,6 +8,8 @@ import org.apache.ibatis.annotations.Param;
 
 import jakarta.validation.Valid;
 import today.meevote.domain.schedule.dto.request.CreatePersonalScheduleDto;
+import today.meevote.domain.schedule.dto.response.GetMyScheduleListDto;
+import today.meevote.domain.schedule.dto.response.GetScheduleCategoryDto;
 
 @Mapper
 public interface ScheduleDao {
@@ -24,5 +27,18 @@ public interface ScheduleDao {
 	public void deleteMemberSchedule(Long scheduleId);
 
 	public void deletePersonalSchedule(Long scheduleId);
+
+	List<GetMyScheduleListDto> getMyScheduleList(
+			@Param("email")
+			String email,
+			@Param("isGroup")
+			Boolean isGroup,
+			@Param("year")
+			String year,
+			@Param("month")
+			String month);
+
+
+	List<GetScheduleCategoryDto> getCategory();
 	
 }
