--- src/main/resources/mapper/scheduleV2Mapper.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-
-
-
-<mapper namespace="today.meevote.domain.schedule.dao.ScheduleV2Dao">
-    <select id="getMyScheduleList" resultType="today.meevote.domain.schedule.dto.response.GetMyScheduleListDto">
-	  SELECT
-	    s.schedule_id AS scheduleId,
-	    s.name,
-	    c.schedule_category_id AS scheduleCategoryId,
-	    c.schedule_category_name AS scheduleCategoryName,
-	    c.schedule_color AS scheduleColor,
-	    s.is_group AS isGroup,
-	    s.duration,
-	    TO_CHAR(s.start_date, 'YYYY-MM-DD HH24:MI') AS startDate,
-	    TO_CHAR(s.end_date, 'YYYY-MM-DD HH24:MI') AS endDate,
-	    s.place_name AS placeName
-	  FROM
-	    member_schedule ms
-	    JOIN schedule s ON ms.schedule_id = s.schedule_id
-	    JOIN schedule_category c ON s.schedule_category_id = c.schedule_category_id
-	  WHERE
-	    ms.email = #{email}
-	    <if test="isGroup != null">
-	      AND s.is_group = #{isGroup}
-	    </if>
-	    AND TO_CHAR(s.start_date, 'YYYY') = #{year}
-	    AND TO_CHAR(s.start_date, 'MM') = #{month}
-	    AND s.end_date IS NOT NULL
-		AND (
-		    (s.is_group = 1
-		         AND s.place_name IS NOT NULL
-		         AND s.place_latitude IS NOT NULL
-		         AND s.place_longitude IS NOT NULL)
-			OR s.is_group = 0)
-	</select>
-
-	<select id="getCategory" resultType="today.meevote.domain.schedule.dto.response.GetScheduleCategoryDto">
-		SELECT
-			schedule_category_id,
-			schedule_category_name,
-			schedule_color
-		FROM schedule_category
-	</select>
-</mapper>
